# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /withdrawals:
    post:
      operationId: initiateWithdrawal
      tags: [withdrawals]
      summary: Initiate a withdrawal
      description: >
        Creates a withdrawal request that locks the requested Winning Coins and
        schedules them for payout to the specified payout account (bank transfer
        or UPI). Requires a verified mobile number and accepted Terms & Conditions.


        Only one active withdrawal (reserved or processing) is permitted per user at a time.


        Idempotent on the idempotencyKey field — retrying with the same key and
        same parameters returns the existing withdrawal (200) rather than creating
        a duplicate.


        Rate limited: 10 requests per 60 seconds per user.
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitiateWithdrawalInput"
      responses:
        "201":
          description: Withdrawal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
        "200":
          description: Idempotent return — same key and params, existing withdrawal returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
        "400":
          description: >
            Validation error, amount below minimum, mobile not verified,
            or Terms & Conditions not accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Payout account not found or does not belong to this user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Active withdrawal already exists, or conflicting idempotency key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Insufficient Winning Coins balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: getWithdrawals
      tags: [withdrawals]
      summary: List the authenticated user's withdrawals
      description: Returns a paginated list of withdrawals, newest first.
      security:
        - sessionCookie: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: before
          in: query
          required: false
          description: Return only withdrawals created before this timestamp.
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: A page of withdrawals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetWithdrawalsResult"
        "400":
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /withdrawals/{id}:
    get:
      operationId: getWithdrawal
      tags: [withdrawals]
      summary: Get a single withdrawal by ID
      description: >
        Returns a withdrawal belonging to the authenticated user.
        Returns 404 (not 403) if the withdrawal belongs to a different user to
        avoid revealing whether a withdrawal ID exists.
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Withdrawal found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
        "400":
          description: Invalid UUID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found or does not belong to this user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /withdrawals/{id}/cancel:
    post:
      operationId: cancelWithdrawal
      tags: [withdrawals]
      summary: Cancel a withdrawal
      description: >
        Cancels a withdrawal and releases the locked Winning Coins back to the
        user's available balance. Only valid while the withdrawal is in the
        `reserved` state (before submission to the payout provider). Idempotent
        if already cancelled.
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Withdrawal cancelled (or was already cancelled)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
        "400":
          description: Invalid UUID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found or does not belong to this user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Withdrawal is not in `reserved` state and cannot be cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sid
      description: Opaque server-side session token issued on signup/login.
  schemas:
...
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    PayoutMethod:
      type: string
      enum: [bank_transfer, upi]
      description: >
        Payout method for a saved account or withdrawal.
        bank_transfer — NEFT/IMPS to a bank account (account number + IFSC code).
        upi           — Instant transfer to a UPI Virtual Payment Address (VPA).
...
    WithdrawalStatus:
      type: string
      enum: [reserved, processing, completed, failed, cancelled]
      description: >
        Lifecycle state of a withdrawal request.
        reserved   — Funds locked. Awaiting background job submission.
        processing — Submitted to payout provider. Awaiting webhook or reconciliation.
        completed  — Provider confirmed payout. Balance permanently debited.
        failed     — Provider rejected or timed out. Funds returned.
        cancelled  — Cancelled by user (only valid from `reserved`). Funds returned.
    Withdrawal:
      type: object
      required:
        - id
        - userId
        - amount
        - status
        - bankAccountId
        - snapshotPayoutMethod
        - snapshotAccountHolderName
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        amount:
          type: integer
          description: Winning Coins to withdraw (1 coin = ₹1). Always positive.
        status:
          $ref: "#/components/schemas/WithdrawalStatus"
        bankAccountId:
          type: string
          format: uuid
          description: >
            ID of the payout account selected at request time. The account may
            be soft-deleted after the withdrawal is created; use the snapshot
            fields for the actual payout details.
        snapshotPayoutMethod:
          $ref: "#/components/schemas/PayoutMethod"
          description: >
            Payout method captured at request time. Determines which snapshot
            fields are populated (bank vs UPI).
        snapshotAccountHolderName:
          type: string
        snapshotBankAccountNumberMasked:
          type: string
          nullable: true
          description: >
            Last 4 digits of the bank account number, prefixed with asterisks.
            Null for UPI withdrawals. Full account number is never returned.
        snapshotBankIfscCode:
          type: string
          nullable: true
          description: IFSC code captured at request time. Null for UPI withdrawals.
        snapshotBankName:
          type: string
          nullable: true
          description: Bank display name captured at request time. Null for UPI withdrawals.
        snapshotUpiId:
          type: string
          nullable: true
          description: >
            UPI VPA captured at request time. Null for bank_transfer withdrawals.
        provider:
          type: string
          nullable: true
          description: >
            Payout provider name (e.g. "payu"). Populated after the withdrawal
            is submitted. Null until then.
        providerReference:
          type: string
          nullable: true
          description: Provider's unique transfer ID. Null until submitted.
        failureReason:
          type: string
          nullable: true
          description: Human-readable failure reason. Null unless status is `failed`.
        cancellationReason:
          type: string
          nullable: true
          description: >
            Cancellation reason. Null unless status is `cancelled`.
        completedAt:
          type: string
          format: date-time
          nullable: true
        failedAt:
          type: string
          format: date-time
          nullable: true
        cancelledAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    InitiateWithdrawalInput:
      type: object
      required: [amount, payoutAccountId, idempotencyKey]
      properties:
        amount:
          type: integer
          minimum: 100
          description: >
            Winning Coins to withdraw. Must be a whole number >= 100 (₹100 minimum).
        payoutAccountId:
          type: string
          format: uuid
          description: >
            ID of a pre-registered, non-deleted payout account (bank transfer or
            UPI) belonging to the authenticated user.
        idempotencyKey:
          type: string
          minLength: 1
          maxLength: 128
          description: >
            Client-generated key. Retrying with the same key and the same amount
            + payoutAccountId returns the existing withdrawal without creating a
            duplicate. Retrying with the same key but different parameters returns
            409.
    GetWithdrawalsResult:
      type: object
      required: [withdrawals]
      properties:
        withdrawals:
          type: array
          items:
            $ref: "#/components/schemas/Withdrawal"